First phase of clean-up post reviews

+++ include/eve/arch/cardinals.hpp
@@ -46,6 +46,6 @@ namespace eve
   //================================================================================================
   // Template inline object for passing index within a wide
   //================================================================================================
-  template<std::uint64_t N>
+  template<std::ptrdiff_t N>
   inline constexpr auto index = std::integral_constant<std::ptrdiff_t,N>{};
 }
